[Add] baseMonster2 에셋 적용
- baseMonster2 에셋 적용
- 원거리 공격에 대한 고민 필요
