+++ Assets/2.Private/LimJH/Scripts/Behavior Tree/TestMonster/Attack_Amber.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 818ac8851a61d264dafcba7a6508e379
+guid: 2845b5ae97b2ea54f9e4718842fc7be9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
